requirements.txt
openpyxl>=3.1.0
pymupdf>=1.24.0
pillow>=10.0.0
pillow-heif>=0.18.0
rich>=13.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
